# apis/openapi.yaml
openapi: 3.0.3
info:
  title: Order Management API
  version: "2.0.0"
  description: >
    The missing upstream Open Host Service for the warehouse-systems
    platform: it owns **Order** and **OrderLine** as first-class, validated
    aggregates. Before this bounded context existed, "an order" was just an
    unowned, unvalidated string — `OrderRef` / `DemandRef` / `Reference` —
    independently reinvented by three different services. Here the OrderId
    is a real identity, and it is the value sent downstream as
    inventory-storage's `demandRef`.


    The lifecycle is intake -> allocation -> release, with cancellation
    legal up to the release boundary — but as of this redesign, a caller
    only ever expresses ONE intent: **place an order**
    (`POST /orders`). Allocation and release are internal saga steps this
    service performs automatically, right after intake, and again on
    `retry-allocation` — they are no longer public REST verbs a caller
    drives by hand. **Allocation** reserves stock per line through
    inventory-storage's published `POST /reservations`; a `409` from that
    service is the business fact "no usable stock", which backorders that
    one line, while any other failure fails the whole allocation pass
    rather than silently recording a backorder. **Release** is no longer a
    synchronous call to any Supplier — it is announced as a fact on the
    enriched `OrderAllocated` / `OrderPartiallyAllocated` Kafka integration
    events (topic `warehouse.order-management.events`), which
    wes-work-planning consumes to enqueue its own work asynchronously. This
    service holds no reservation or work-unit state of its own, only the
    references it needs to revoke.


    The order-level `status` is always DERIVED from the line statuses and
    is never stored, so it cannot drift out of sync with the lines it
    summarises.


    Out of this service's boundary: it does not hold inventory, pick, pack,
    ship, or plan work (those are inventory-storage, wes-work-planning and
    fulfillment-execution), and it does not model a local Reservation or
    WorkUnit aggregate. Error responses across this API use RFC 7807
    (`application/problem+json`) Problem Details.
  contact:
    name: Order Management (Generic/Supporting Bounded Context)
servers:
  - url: http://localhost:8080
    description: Local development
# Fleet REST identity (ADR-0011, adopting warehouse-ops-agent ADR 0005):
# every operation except /healthz requires a static bearer key.
# GET/HEAD/OPTIONS need the read scope; every other method needs read-write.
security:
  - bearerAuth: []
tags:
  - name: Orders
    description: >
      Order intake and reads. An order is received with its lines in
      `Received` status, and this service immediately attempts
      allocation-then-release in the SAME call — a caller expresses intent
      by placing the order, not by separately driving internal saga steps.
  - name: Allocation
    description: >
      Automatic allocation of stock per line through inventory-storage,
      folded into `POST /orders` and `POST /orders/{id}/retry-allocation`.
      A `409` from that Supplier backorders a single line (a business
      fact); any other failure fails the whole allocation pass
      (fail-closed on ambiguity). A backordered line returns to
      `Allocated` only via retry-allocation, which also re-attempts
      release for any lines it newly clears.
  - name: Health
    description: Liveness probe.

paths:
  /healthz:
    get:
      operationId: getHealthz
      summary: Liveness probe
      description: >
        Returns 200 with a static status body if the process is up. Does
        not check downstream dependencies (Postgres, inventory-storage,
        wes-work-planning) — it is a liveness check, not a readiness check.
      tags: [Health]
      security: []
      responses:
        "200":
          description: Service is alive.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                example:
                  status: ok

  /orders:
    post:
      operationId: receiveOrder
      summary: Receive a new order — allocates and releases automatically
      description: >
        Intake — the ReceiveOrder use case. Validates every requested line
        (SKU non-empty, quantity > 0), mints an `OrderId`, and persists the
        order in `Received` status, publishing `OrderReceived`
        unconditionally.


        Immediately afterward, in the SAME call, this service attempts
        allocation-then-release: it reserves stock per line via
        inventory-storage's `POST /reservations` (see the Allocation tag),
        and for every line that ends up `Allocated` and eligible under BR3
        (see below), releases it — announced via the enriched
        `OrderAllocated` / `OrderPartiallyAllocated` Kafka events, never a
        second synchronous call. If that implicit allocation attempt hits
        a hard (non-business) failure, this endpoint still returns `201`:
        the order genuinely was received, and a caller expressing "place
        an order" must never be told that failed just because an internal,
        best-effort next step could not complete. The returned body
        reflects whatever the allocation pass actually achieved.


        `allowPartialShipment` defaults to `false` — ship-complete (BR3).
        Every line's `pathId` is always the internal default (`pick`) — a
        caller placing an order has no business supplying
        wes-work-planning's process-path vocabulary at intake; `pathId` is
        visible on every line in the response, but not settable on the
        request.
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiveOrderRequest"
            example:
              allowPartialShipment: false
              lines:
                - sku: SKU-1
                  quantity: 2
                - sku: SKU-2
                  quantity: 1
                  giftWrap: true
      responses:
        "201":
          description: >
            Order received. Reflects whatever the immediate
            allocation-then-release attempt achieved — `Received`,
            `Backordered`, `PartiallyAllocated`/`PartiallyReleased`,
            `Allocated`, or `Released` — a caller inspects `status`/line
            `status` for the outcome rather than assuming `Received`.
          headers:
            Location:
              schema:
                type: string
              description: Canonical URI of the created order.
              example: /orders/ord-a1b2c3d4-0000-0000-0000-000000000001
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
              example:
                id: ord-a1b2c3d4-0000-0000-0000-000000000001
                status: Released
                allowPartialShipment: false
                promiseDate: "2026-08-27T09:00:00Z"
                lines:
                  - lineNo: 1
                    sku: SKU-1
                    quantity: 2
                    pathId: pick
                    giftWrap: false
                    status: Released
                    reservationId: res-a1b2c3d4-0000-0000-0000-000000000009
                  - lineNo: 2
                    sku: SKU-2
                    quantity: 1
                    pathId: pick
                    giftWrap: true
                    status: Released
                    reservationId: res-a1b2c3d4-0000-0000-0000-000000000010
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          description: A line quantity is not greater than zero.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.order-management.warehouse-systems.dev/non-positive-quantity
                title: Quantity must be greater than zero
                status: 422
                detail: quantity must be greater than zero
                instance: /orders
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InsufficientScope"
        "500":
          $ref: "#/components/responses/InternalError"

  /orders/{id}:
    get:
      operationId: getOrder
      summary: Read an order's current state
      description: >
        The GetOrder use case. Returns the order with every line's current
        status, its reservation reference where allocated, and the promise
        date once allocation has computed one. The order-level `status` in
        this response is derived from the line statuses at read time — it
        is not a stored field.
      tags: [Orders]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: The order's current state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
              example:
                id: ord-a1b2c3d4-0000-0000-0000-000000000001
                status: Allocated
                allowPartialShipment: false
                promiseDate: "2026-08-27T09:00:00Z"
                lines:
                  - lineNo: 1
                    sku: SKU-1
                    quantity: 2
                    pathId: pick
                    giftWrap: false
                    status: Allocated
                    reservationId: res-a1b2c3d4-0000-0000-0000-000000000009
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/OrderNotFound"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InsufficientScope"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: cancelOrder
      summary: Cancel an order and revoke its reservations
      description: >
        The CancelOrder use case. Revokes every allocated line's
        reservation on inventory-storage via
        `DELETE /reservations/{id}`, then cancels every line. Publishes
        `OrderCancelled`.


        **BR6 — the cancellation boundary:** this is legal ONLY while no
        line has reached `Released`. The boundary is checked BEFORE any
        reservation is revoked, so a rejected cancellation leaves
        inventory-storage completely untouched. Once any line is released,
        v1 does NOT claw back the released work — a deliberate, documented
        gap (see ADR 0004), not an oversight.


        A failed revoke fails the whole cancellation and leaves the order
        untouched, so retrying is safe.
      tags: [Orders]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "204":
          description: Order cancelled; every allocated reservation was revoked.
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/OrderNotFound"
        "409":
          description: At least one line has already been released (BR6).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.order-management.warehouse-systems.dev/order-already-released
                title: Order already has released lines and can no longer be cancelled
                status: 409
                detail: order already has released lines and can no longer be cancelled
                instance: /orders/ord-a1b2c3d4-0000-0000-0000-000000000001
        "503":
          $ref: "#/components/responses/DownstreamUnavailable"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InsufficientScope"
        "500":
          $ref: "#/components/responses/InternalError"

  /orders/{id}/retry-allocation:
    post:
      operationId: retryAllocation
      summary: Re-attempt allocation for backordered lines, then release
      description: >
        The RetryAllocation use case. Re-attempts allocation for the
        order's `Backordered` lines and ONLY those lines — it is the single
        sanctioned route from `Backordered` back to `Allocated`, and it is
        what unblocks a ship-complete order held back by BR3. UNLIKE the
        implicit allocation attempt folded into `POST /orders`, this is an
        explicit, on-purpose operator/recovery action: a hard (non-409)
        failure here fails the call with `503` rather than being swallowed
        — invoking this endpoint IS asking for allocation right now.


        A line still short of stock stays `Backordered` (no error). The
        same fail-closed rule as intake's implicit allocation applies to
        any non-409 failure. An order with no backordered line at all is
        rejected with `409`: retrying an order that is not backordered is
        a caller mistake, not a no-op.


        On success, this ALSO attempts release in the same call — the same
        allocation-then-release flow `POST /orders` uses — announced via
        the enriched `OrderAllocated` / `OrderPartiallyAllocated` Kafka
        events rather than any synchronous call to a Supplier.
      tags: [Allocation]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: The order after the retry-then-release pass.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
              example:
                id: ord-a1b2c3d4-0000-0000-0000-000000000001
                status: Released
                allowPartialShipment: false
                promiseDate: "2026-08-27T09:00:00Z"
                lines:
                  - lineNo: 1
                    sku: SKU-1
                    quantity: 2
                    pathId: pick
                    giftWrap: false
                    status: Released
                    reservationId: res-a1b2c3d4-0000-0000-0000-000000000009
                  - lineNo: 2
                    sku: SKU-2
                    quantity: 1
                    pathId: pick
                    giftWrap: true
                    status: Released
                    reservationId: res-a1b2c3d4-0000-0000-0000-000000000010
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/OrderNotFound"
        "409":
          description: This order has no backordered line to retry.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.order-management.warehouse-systems.dev/no-backordered-lines
                title: Order has no backordered lines to retry
                status: 409
                detail: order has no backordered lines to retry
                instance: /orders/ord-a1b2c3d4-0000-0000-0000-000000000001/retry-allocation
        "503":
          $ref: "#/components/responses/DownstreamUnavailable"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InsufficientScope"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Static API key with a `read` or `read-write` scope (ADR-0011,
        fleet ADR 0005). Keys are issued out of band from a Kubernetes
        Secret (`API_READ_KEY` / `API_READWRITE_KEY`); there is no IdP.
        `GET`/`HEAD`/`OPTIONS` require `read`; every other method requires
        `read-write`. `/healthz` is open.

  parameters:
    OrderId:
      name: id
      in: path
      required: true
      description: The order's identifier, as returned by receiveOrder.
      schema:
        type: string
        minLength: 1
      example: ord-a1b2c3d4-0000-0000-0000-000000000001

  responses:
    BadRequest:
      description: The request body or path parameter is malformed or invalid.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.order-management.warehouse-systems.dev/empty-sku
            title: SKU must not be empty
            status: 400
            detail: sku must not be empty
            instance: /orders

    OrderNotFound:
      description: No order has this id.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.order-management.warehouse-systems.dev/order-not-found
            title: Order not found
            status: 404
            detail: order not found
            instance: /orders/ord-does-not-exist

    DownstreamUnavailable:
      description: >
        inventory-storage could not be reached, answered ambiguously, or
        is wired in permissive (no-op) mode. Never papered over with a
        2xx: allocating real stock must not appear to succeed when it did
        not.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.order-management.warehouse-systems.dev/downstream-not-configured
            title: A downstream service is running in permissive (no-op) mode
            status: 503
            detail: downstream client is running in permissive (no-op) mode and cannot perform this operation
            instance: /orders/ord-a1b2c3d4-0000-0000-0000-000000000001/retry-allocation

    InternalError:
      description: An unexpected internal error occurred.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.order-management.warehouse-systems.dev/internal-error
            title: An unexpected internal error occurred
            status: 500
            detail: unexpected adapter failure
            instance: /orders

    Unauthenticated:
      description: >
        No bearer credential, or an unknown one. Carries a
        `WWW-Authenticate: Bearer` challenge. Only emitted when the service
        runs with `AUTH_MODE=enforce` (ADR-0011).
      headers:
        WWW-Authenticate:
          description: The bearer challenge.
          schema:
            type: string
            example: Bearer realm="warehouse-systems"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.order-management.warehouse-systems.dev/unauthenticated
            title: Unauthorized
            status: 401
            detail: Missing or invalid bearer credential.
            instance: /orders

    InsufficientScope:
      description: >
        The credential is valid but grants only the `read` scope, and this
        operation requires `read-write`. Only emitted under
        `AUTH_MODE=enforce` (ADR-0011).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.order-management.warehouse-systems.dev/insufficient-scope
            title: Forbidden
            status: 403
            detail: This credential does not grant the read-write scope.
            instance: /orders

  schemas:
    Problem:
      type: object
      description: >
        RFC 7807 (Problem Details for HTTP APIs) error shape, used for
        every non-2xx response across this service — the same shape the
        other five services in this platform emit.
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
          description: >
            A URI identifying the error category (does not need to
            resolve). Unique per distinct error category in this service.
          example: https://errors.order-management.warehouse-systems.dev/order-not-found
        title:
          type: string
          description: Fixed, human-readable summary of the error category.
          example: Order not found
        status:
          type: integer
          format: int32
          description: Duplicates the HTTP status code as an integer (RFC 7807 requirement).
          example: 404
        detail:
          type: string
          description: The specific error message for this occurrence.
          example: order not found
        instance:
          type: string
          description: The request path that produced this error.
          example: /orders/ord-does-not-exist

    ReceiveOrderRequest:
      type: object
      required: [lines]
      properties:
        lines:
          type: array
          minItems: 1
          description: The requested lines. An order must have at least one.
          items:
            $ref: "#/components/schemas/ReceiveOrderLineRequest"
        allowPartialShipment:
          type: boolean
          default: false
          description: >
            When false (the default — ship-complete, BR3), a single
            backordered line holds the WHOLE order back from release until
            retry-allocation clears it. When true, allocated lines are
            independently eligible for release.
          example: false

    ReceiveOrderLineRequest:
      type: object
      required: [sku, quantity]
      properties:
        sku:
          type: string
          minLength: 1
          description: The stock keeping unit ordered. Must be non-empty.
          example: SKU-1
        quantity:
          type: integer
          format: int32
          minimum: 1
          description: Quantity ordered. Must be greater than zero.
          example: 2
        giftWrap:
          type: boolean
          default: false
          description: >
            OPTIONAL. Whether the requester asked the warehouse to produce
            a gift package for this line. Passed straight through onto the
            enriched `OrderAllocated`/`OrderPartiallyAllocated` Kafka
            event's `lines[].gift_wrap` field — a caller-stated
            characteristic, never looked up.
          example: true

    OrderStatus:
      type: string
      description: >
        The order-level status, ALWAYS derived from the line statuses and
        never stored, so it cannot drift out of sync with the lines it
        summarises. `Received` -> `Allocated` | `PartiallyAllocated` |
        `Backordered` -> `Released` | `PartiallyReleased`, with `Cancelled`
        reachable only from a pre-release state (BR6).
      enum:
        - Received
        - Allocated
        - PartiallyAllocated
        - Backordered
        - Released
        - PartiallyReleased
        - Cancelled
      example: Allocated

    OrderLineStatus:
      type: string
      description: >
        The per-line status. `Pending` -> `Allocated` (inventory-storage
        reserved) or `Backordered` (inventory-storage answered 409);
        `Backordered` -> `Allocated` ONLY via retry-allocation;
        `Allocated` -> `Released` (wes-work-planning accepted the work);
        any pre-release state -> `Cancelled`.
      enum: [Pending, Allocated, Backordered, Released, Cancelled]
      example: Allocated

    OrderLine:
      type: object
      required: [lineNo, sku, quantity, pathId, giftWrap, status]
      properties:
        lineNo:
          type: integer
          format: int32
          description: The line's 1-based position within its order.
          example: 1
        sku:
          type: string
          example: SKU-1
        quantity:
          type: integer
          format: int32
          example: 2
        pathId:
          type: string
          description: >
            The process path assigned to this line, ALWAYS the internal
            default (`pick`) — a caller placing an order has no business
            supplying wes-work-planning's process-path vocabulary at
            intake, so this is read-only: visible in every response, never
            settable on the request.
          example: pick
        giftWrap:
          type: boolean
          example: false
        status:
          $ref: "#/components/schemas/OrderLineStatus"
        reservationId:
          type: string
          description: >
            inventory-storage's reservation id, present only once the line
            is allocated. A reference only — this context models no local
            Reservation aggregate, because inventory-storage remains the
            sole owner of reservation state. Omitted (rather than empty)
            when the line is not allocated.
          example: res-a1b2c3d4-0000-0000-0000-000000000009

    Order:
      type: object
      required: [id, status, allowPartialShipment, lines]
      properties:
        id:
          type: string
          description: The order's identity — this bounded context's contribution to the platform.
          example: ord-a1b2c3d4-0000-0000-0000-000000000001
        status:
          $ref: "#/components/schemas/OrderStatus"
        allowPartialShipment:
          type: boolean
          description: False means ship-complete (BR3), the default.
          example: false
        promiseDate:
          type: string
          format: date-time
          description: >
            The date this order is promised for, RFC 3339. Computed at
            allocation time by the domain lead-time policy as "now plus the
            slowest lead time among the allocated lines' process paths".
            Absent until at least one line is allocated. v1 uses a
            configurable static lead time, not a live carrier rate — see
            the README's "Deferred (v1)" section.
          example: "2026-08-27T09:00:00Z"
        lines:
          type: array
          items:
            $ref: "#/components/schemas/OrderLine"